Backend ready

+++ admin/sql/install.mysql.utf8.sql
@@ -5,9 +5,9 @@ CREATE TABLE `#__downloads` (
   `name` VARCHAR(225) NOT NULL,
   `email` VARCHAR(225) NOT NULL,
   `file` VARCHAR(225) NOT NULL,
-  `count` tinyint(11) NOT NULL,
-  `created_at` tinyint(11) NOT NULL,
-  `updated_at` tinyint(11) NOT NULL,
+  `count` int(11) NOT NULL,
+  `created_at` int(11) NOT NULL,
+  `updated_at` int(11) NOT NULL,
   PRIMARY KEY (`id`)
 )
   ENGINE =MyISAM
